+++ src/workbench/studio-graphs/apolloStudioGraphsTreeDataProvider.ts
@@ -2,6 +2,7 @@ import path from 'path';
 import * as vscode from 'vscode';
 import { getAccountGraphs, getGraphSchemasByVariant, getUserMemberships } from '../../studio-gql/graphClient';
 import { GetGraphSchemas_service_implementingServices_FederatedImplementingServices } from '../../studio-gql/types/GetGraphSchemas';
+import { PreloadedWorkbenchTopLevel } from '../local-workbench-files/preLoadedTreeItems';
 
 
 export class ApolloStudioGraphsTreeDataProvider implements vscode.TreeDataProvider<vscode.TreeItem> {
@@ -105,10 +106,9 @@ export class ApolloStudioGraphsTreeDataProvider implements vscode.TreeDataProvid
                 if (response === "Login")
                     vscode.commands.executeCommand("extension.enterStudioApiKey");
             });
-            // let response = await vscode.window.showInformationMessage('No user api key was found.', "Login");
-            // if (response === "Login")
-            //     vscode.commands.executeCommand("extension.enterStudioApiKey");
         }
+
+        items.push(new PreloadedWorkbenchTopLevel());
 
         return items;
     }